[Queens-only] Ensure epmd is spawned by systemd before redeploy
On (re)deploy, we forcibly restart epmd and rabbit to ensure any
upgrade of erlang or rabbitmq can be taken into account before
puppet runs. But there is currently no guarantee as to which
one of systemd or rabbitmq will restart epmd.

This is an improvement over I6f486b0b70f19d8b4916ef500675c0739939e060,
to selectively kill epmd and ensure the systemd service will always
restart it, like it is supposed to.

Change-Id: I82b419cfe4585a3f185687b8cc0a19bb534f7536
Closes-Bug: #1823003