Quote controller_public_host_wrapped
If this is an IPv6 address, it currently will be set to something like
controller_public_host_wrapped: [2001:db8::2]

The yaml parser interprets this as a list and tries to part the IPv6
address as a hash. To avoid this, we need to wrap the value in quotes.

Change-Id: I9e5d5e6cd6873431a17280a8a8c8ae5af45163c4
Closes-Bug: #1833429
(cherry picked from commit e4cc043a1c945dc527f23897df4f06f9fb5ce2b8)