Stop hardcoding baremetal API version in stackrc
Long time ago it was required because ironicclient defaulted to
an ancient version. It was changed back in Queens (ironicclient 2.0),
so we can drop it now to avoid confusion.

This is the same as change Icea0bdf6d5dcdd81ce9c34be7af8a241da0861bc,
but adapter to instack-undercloud.

Change-Id: I106f5dc0ed4c3a3b67d8fc94c69b2b2f1d8fc547
Closes-Bug: #1789392