Remove nova-cert service during FFU
As nova-cert service doesn't exist on fresh queens
deployment, but it is enabled in newton.
During FFU from newton to queens, it is shown in
nova service list in ocata, pike and queens as well.

And since ocata and pike are EOL, we need to remove
it in queens during FFU from undercloud.

Change-Id: I1da2599ae6b0c5e860ce3a71fe6359ec649e8958